fix:添加对x86兼容
[Mantis ID]:
[Cus ID]:
[What]:
[Why]:
[How]:
[Test Method]:
[Red Flag]:

+++ WindowsTest/Form1.cs
@@ -16,7 +16,7 @@ namespace WindowsTest
     public partial class Form1 : Form
     {
         IntPtr mMonitorList;
-        private ulong volumeValue = 0;
+        private UInt32 volumeValue = 0;
         public class ComboboxItem
         {
             public string Text { get; set; }
